Add response-level identity map
When response rows reference the same node or relationship, the response
currently returns new object instances for each one. Since they're all
part of the same response, it's reasonable to assume they can be the
same object, so this identity-map prototype makes that happen, saving a
bit of RAM and possibly a bit of CPU time for the conversion from hash
to ActiveNode/ActiveRel.