Check Neo4j server responsiveness for CI
This PR is intended to make CI builds faster without jumping the gun on the server by looping until the server responds. This will also guard against the possibility of the Neo4j server taking > 2 minutes to accept connections.

It is throttled to check only once per second so it doesn't waste CPU resources spinning up processes that it could be spending on spinning up the Neo4j server.